techDebt(buRefactorTraining): Rename uncheckedWorker to noAssociatedWorker
